complete Z2

+++ v12z2/yl_board.h
@@ -3,9 +3,11 @@
 
 #include <wiringPi.h>
 #include <wiringPiI2C.h>
+#include <stdlib.h>
+#include <stdio.h>
 
 void yl_init(int &fd);
 
-void readYL(int &fd, double &adc, double &t, double &light);
+void readYL(int &fd, int &adc, double &t, int &light);
 
 #endif // YL_BOARD_H
